SDN-1233: Adjust packet socket binding
Between the time when a packet socket is opened and until it's bound to
an interface, it receives packets meant for other interfaces as well. When
polling is started on this file descriptor, it triggers events for these
existing packets too, which is seen as ifname mismatch error. Doing a
socket bind immediately after opening the socket reduces the probability
and any such packets and can be safely ignored from server perspective.

Signed-off-by: Reshma Sreekumar <reshma.sreekumar@ionos.com>