DHCPv6: remove append 0xff on options pack
This fixes DHCPv6 support for Windows and Debian-based Guests

It looks like adding a byte 0xff to the end of the list of dhcpv6
options is a Wrong Thing. RFC-3315 does not mention anything of the
kind, and formally speaking, such byte would be interpreted as an
"unparseable option", because minimal size of an option is 32 bits.
From experience, some DHCPv6 clients are more lenient than others, and
such messages from the server are successfully parsed by, for instance,
OpenWRT. Other clients ignore these messages as invalid.

Jira: SDN-1455
Signed-off-by: Mikhail Sennikovsky <mikhail.sennikovskii@ionos.com>
Signed-off-by: Eugene Crosser <evgenii.cherkashin@ionos.com>